Move the ControlledCodeMirror and EditorWithAutocomplete implementation into a shared folder
